+++ src/circuit/circuit_vanilla.rs
@@ -803,7 +803,7 @@ mod tests {
             SingleVerifier, VerifyingKey, K,
         },
         keys::SpendValidatingKey,
-        note::{AssetBase, Note, NoteVersion, Rho},
+        note::{Note, NoteVersion, Rho},
         tree::MerklePath,
         value::{ValueCommitTrapdoor, ValueCommitment},
     };
@@ -864,7 +864,7 @@ mod tests {
 
         let value = spent_note.value() - output_note.value();
         let rcv = ValueCommitTrapdoor::random(&mut rng);
-        let cv_net = ValueCommitment::derive(value, rcv.clone(), AssetBase::zatoshi());
+        let cv_net = ValueCommitment::derive(value, rcv.clone());
 
         let path = MerklePath::dummy(&mut rng);
         let anchor = path.root(spent_note.commitment().into());